flask_backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
google-auth==2.22.0
PyJWT==2.8.0
python-jose==3.3.0
supabase==1.0.3